Refactor child components into seperate containers.
